requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Flask-cors
PyMySQL
WTForms
oss2
black
isort